Crate jsonc_parser

Modules

ast

common

errors

tokens

Structs

CollectOptions

Options for collecting comments and tokens.

JsonArray

A JSON array.

JsonObject

A JSON object.

ParseOptions

Options for parsing.

ParseResult

Result of parsing the text.

Scanner

Converts text into a stream of tokens.

Enums

JsonValue

A JSON value.

Functions

parse_to_ast

Parses a string containing JSONC to an AST with comments and tokens.

parse_to_serde_value

Parses a string containing JSONC to a `serde_json::Value.

parse_to_value

Parses a string containing JSONC to a JsonValue.

Type Aliases

CommentMap

Map where the comments are stored in collections where the key is the previous token end or start of file or next token start or end of the file.