Module jsonc_parser::ast

Structs

Array

Represents an array that may contain elements (ex. [], [5, 6]).

BooleanLit

Represents a boolean (ex. true or false).

CommentBlock

Represents a comment block (ex. /* my comment */).

CommentLine

Represents a comment line (ex. // my comment).

NullKeyword

Represents the null keyword (ex. null).

NumberLit

Represents a number (ex. 123, 99.99, -1.2e+2).

Object

Represents an object that may contain properties (ex. {}, { "prop": 4 }).

ObjectProp

Represents an object property (ex. "prop": []).

StringLit

Node surrounded in double quotes (ex. "my string").

WordLit

A string that’s not in quotes. Usually the appearance of this would be a parsing error.

Enums

Comment

JSONC comment.

CommentKind

Kind of JSONC comment.

Node

Node that can appear in the AST.

NodeKind

Kind of AST node.

ObjectPropName

Represents an object property name that may or may not be in quotes.

Value

JSON value.