jsonc_parser::ast

Struct Object

pub struct Object<'a> {
    pub range: Range,
    pub properties: Vec<ObjectProp<'a>>,
}

Represents an object that may contain properties (ex. {}, { "prop": 4 }).

Fields
range: Rangeproperties: Vec<ObjectProp<'a>>
Implementations

impl<'a> Object<'a>

pub fn get(&self, name: &str) -> Option<&ObjectProp<'a>>

Gets a property value in the object by its name.

pub fn get_string(&self, name: &str) -> Option<&StringLit<'a>>

Gets a string property value from the object by name. Returns None when not a string or it doesn’t exist.

pub fn get_number(&self, name: &str) -> Option<&NumberLit<'a>>

Gets a number property value from the object by name. Returns None when not a number or it doesn’t exist.

pub fn get_boolean(&self, name: &str) -> Option<&BooleanLit>

Gets a boolean property value from the object by name. Returns None when not a boolean or it doesn’t exist.

pub fn get_object(&self, name: &str) -> Option<&Object<'a>>

Gets an object property value from the object by name. Returns None when not an object or it doesn’t exist.

pub fn get_array(&self, name: &str) -> Option<&Array<'a>>

Gets an array property value from the object by name. Returns None when not an array or it doesn’t exist.

pub fn take(&mut self, name: &str) -> Option<ObjectProp<'a>>

Takes a value from the object by name. Returns None when it doesn’t exist.

pub fn take_string(&mut self, name: &str) -> Option<StringLit<'a>>

Takes a string property value from the object by name. Returns None when not a string or it doesn’t exist.

pub fn take_number(&mut self, name: &str) -> Option<NumberLit<'a>>

Takes a number property value from the object by name. Returns None when not a number or it doesn’t exist.

pub fn take_boolean(&mut self, name: &str) -> Option<BooleanLit>

Takes a boolean property value from the object by name. Returns None when not a boolean or it doesn’t exist.

pub fn take_object(&mut self, name: &str) -> Option<Object<'a>>

Takes an object property value from the object by name. Returns None when not an object or it doesn’t exist.

pub fn take_array(&mut self, name: &str) -> Option<Array<'a>>

Takes an array property value from the object by name. Returns None when not an array or it doesn’t exist.