jsonc_parser::common
Trait Ranged

pub trait Ranged {
    // Required method
    fn range(&self) -> &Range;

    // Provided methods
    fn start(&self) -> usize { ... }
    fn end(&self) -> usize { ... }
    fn text<'a>(&self, text: &'a str) -> &'a str { ... }
    fn width(&self) -> usize { ... }
}

Represents an object that has a range in the text.

Required Methods

fn range(&self) -> &Range

Gets the range.

Provided Methods

fn start(&self) -> usize

Gets the byte index of the first character in the text.

fn end(&self) -> usize

Gets the byte index after the last character in the text.

fn text<'a>(&self, text: &'a str) -> &'a str

Gets the text from the provided string.

fn width(&self) -> usize

Gets the end byte index minus the start byte index of the range.