jsonc_parser

Struct JsonObject

pub struct JsonObject<'a>(/* private fields */);

A JSON object.

Implementations

impl<'a> JsonObject<'a>

pub fn new(inner: Map<String, JsonValue<'a>>) -> JsonObject<'a>

Creates a new JsonObject.

pub fn with_capacity(capacity: usize) -> JsonObject<'a>

Creates a new JsonObject with the specified capacity.

pub fn take_inner(self) -> Map<String, JsonValue<'a>>

Drops the object returning the inner map.

pub fn len(&self) -> usize

Gets the number of properties.

pub fn is_empty(&self) -> bool

Gets if there are no properties.

pub fn get(&self, name: &str) -> Option<&JsonValue<'a>>

Gets a value in the object by its name.

pub fn get_string(&self, name: &str) -> Option<&Cow<'a, str>>

Gets a string property value from the object by name. Returns None when not a string or it doesn’t exist.

pub fn get_number(&self, name: &str) -> Option<&'a str>

Gets a number property value from the object by name. Returns None when not a number or it doesn’t exist.

pub fn get_boolean(&self, name: &str) -> Option<bool>

Gets a boolean property value from the object by name. Returns None when not a boolean or it doesn’t exist.

pub fn get_object(&self, name: &str) -> Option<&JsonObject<'a>>

Gets an object property value from the object by name. Returns None when not an object or it doesn’t exist.

pub fn get_array(&self, name: &str) -> Option<&JsonArray<'a>>

Gets an array property value from the object by name. Returns None when not an array or it doesn’t exist.

pub fn take(&mut self, name: &str) -> Option<JsonValue<'a>>

Takes a value from the object by name. Returns None when it doesn’t exist.

pub fn take_string(&mut self, name: &str) -> Option<Cow<'a, str>>

Takes a string property value from the object by name. Returns None when not a string or it doesn’t exist.

pub fn take_number(&mut self, name: &str) -> Option<&'a str>

Takes a number property value from the object by name. Returns None when not a number or it doesn’t exist.

pub fn take_boolean(&mut self, name: &str) -> Option<bool>

Takes a boolean property value from the object by name. Returns None when not a boolean or it doesn’t exist.

pub fn take_object(&mut self, name: &str) -> Option<JsonObject<'a>>

Takes an object property value from the object by name. Returns None when not an object or it doesn’t exist.

pub fn take_array(&mut self, name: &str) -> Option<JsonArray<'a>>

Takes an array property value from the object by name. Returns None when not an array or it doesn’t exist.