jsonc_parser
Type Alias CommentMap

pub type CommentMap<'a> = HashMap<usize, Rc<Vec<Comment<'a>>>>;

Map where the comments are stored in collections where the key is the previous token end or start of file or next token start or end of the file.

Aliased Type

struct CommentMap<'a> { /* private fields */ }